requirements.txt
streamlit==1.32.0
anthropic==0.18.1
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.2.0
python-dotenv==1.0.1